+++ services/ui-src/src/components/pages/Dashboard/DashboardTable.tsx
@@ -1,6 +1,7 @@
 // components
 import { Button, Image, Td, Tr } from "@chakra-ui/react";
-import { AdminActionButtons, Table } from "components";
+import { Spinner } from "@cmsgov/design-system";
+import { Table } from "components";
 // utils
 import { AnyObject, ReportMetadataShape } from "types";
 import { convertDateUtcToEt } from "utils";
@@ -121,6 +122,66 @@ const DateFields = ({ report, reportType }: DateFieldProps) => {
 interface DateFieldProps {
   report: ReportMetadataShape;
   reportType: string;
+}
+
+const AdminActionButtons = ({
+  report,
+  reportType,
+  reportId,
+  archiveReport,
+  archiving,
+  unlocking,
+  unlockReport,
+  sxOverride,
+}: AdminActionButtonProps) => {
+  return (
+    <>
+      {reportType === "MLR" && (
+        <Td>
+          <Button
+            variant="link"
+            disabled={report.locked}
+            sx={sxOverride.adminActionButton}
+            onClick={() => unlockReport!(report)}
+          >
+            {unlocking && reportId === report.id ? (
+              <Spinner size="small" />
+            ) : (
+              "Unlock"
+            )}
+          </Button>
+        </Td>
+      )}
+
+      <Td>
+        <Button
+          variant="link"
+          sx={sxOverride.adminActionButton}
+          onClick={() => archiveReport(report)}
+        >
+          {archiving && reportId === report.id ? (
+            <Spinner size="small" />
+          ) : report?.archived ? (
+            "Unarchive"
+          ) : (
+            "Archive"
+          )}
+        </Button>
+      </Td>
+    </>
+  );
+};
+
+interface AdminActionButtonProps {
+  report: ReportMetadataShape;
+  reportType: string;
+  reportId: string | undefined;
+  archiveReport: Function;
+  archiving: boolean;
+  unlocking?: boolean;
+  unlockReport?: Function;
+  mobile?: boolean;
+  sxOverride: AnyObject;
 }
 
 const sx = {
