+++ services/ui-src/src/utils/testing/setupJest.tsx
@@ -596,7 +596,7 @@ export const mockReportsByState = [
 
 export const mockReportMethods = {
   archiveReport: jest.fn(),
-  lockReport: jest.fn(),
+  unlockReport: jest.fn(),
   fetchReport: jest.fn(),
   fetchReportsByState: jest.fn(),
   createReport: jest.fn(),
